add discord support
