+++ .env.example
@@ -2,6 +2,7 @@
 
 SLACK_TOKEN=
 SLACK_SIGNING_SECRET=
+DISCORD_TOKEN=
 
 # Optional
