+++ GLITCH_README.md
@@ -6,7 +6,7 @@ If you're as concious as I am about the use of language in your Slack workspace,
 
 With the existing [guys-bot-for-slack](https://glitch.com/~guys-bot-for-slack) needing some updates to work with the latest flavour of Slack API, I chose to recreate it using the most basic implementation of the Slack SDKs.
 
-![Screenshot of the bot responding to the message](https://github.com/lukeocodes/guys-bot/raw/master/screenshot.gif "Screenshot of the bot responding to the message")
+![Screenshot of the bot responding to the message](https://github.com/lukeocodes/guys-bot/raw/master/screenshot-slack.gif "Screenshot of the bot responding to the message")
 
 The bot responds with an ephemeral message, not seen by anyone but the user who originally messaged.
