+++ package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@slack/events-api": "^2.3.3",
     "@slack/web-api": "^5.10.0",
+    "discord.js": "^12.3.1",
     "dotenv": "^8.2.0"
   },
   "devDependencies": {
